Remove a couple of superfluous size_t() casts.

+++ src/coupling_ramp.cpp
@@ -39,7 +39,7 @@ void CouplingRamp::init( Configf& configf ) {
     exit(EXIT_FAILURE);
   }
 
-  if( (tempt.size() != size_t(pairs)) || (tempn.size() != size_t(pairs))) {
+  if( (tempt.size() != pairs) || (tempn.size() != pairs)) {
     cerr<<"The length of either *nus* or *timepoints* does not match the number specified in *pairs*" <<endl;
     exit(EXIT_FAILURE);
   }
